C99 guarantees the return of vsnprintf(3)
This function wasn't present in C89.  It was added in C99, where it
returns the size.

SUSv2 (year 1997) already had vsnprintf(3), and it specified already
a return value.  BTW, there was a historic difference that was later
fixed in C99 and POSIX.1-2001, which is mentioned in the HISTORY section
in the manual page in the Linux man-pages, but this detail shouldn't
affect this patch.

snprintf(3):
HISTORY
	[...]

       snprintf()
       vsnprintf()
              SUSv2, C99, POSIX.1‐2001.

              Concerning the return value of snprintf(), SUSv2 and C99
              contradict  each  other:  when snprintf() is called with
              size=0 then SUSv2 stipulates an unspecified return value
              less than 1, while C99 allows str to  be  NULL  in  this
              case, and gives the return value (as always) as the num‐
              ber  of  characters that would have been written in case
              the output string has been large  enough.   POSIX.1‐2001
              and  later  align their specification of snprintf() with
              C99.

Signed-off-by: Alejandro Colomar <alx@nginx.com>